refactor(review): 리뷰 이미지파일을 이미지파일에서 관리하도록 변경 - 컨트롤러에서 하던 유효성검사는 도메인 로직으로 변경함에 따라 테스트 로직 삭제

+++ src/main/java/hanghae/review/shop/product/domain/Product.java
@@ -14,7 +14,7 @@ public class Product {
     private Long reviewCount;
     private Float totalScore;
 
-    public void updateReviewData(Float score) {
+    public void updateReviewData(Integer score) {
         this.reviewCount += 1;
         this.totalScore += score;
     }
