+++ src/main/java/hanghae/review/shop/product/event/ProductIncreaseEvent.java
@@ -2,7 +2,7 @@ package hanghae.review.shop.product.event;
 
 public record ProductIncreaseEvent(
         Long productId,
-        Float score
+        Integer score
 
 ) {
 }
